+++ clang-tools-extra/clang-tidy/misrac/MisracTidyModule.cpp
@@ -18,6 +18,7 @@
 #include "Rule10p8Check.h"
 #include "Rule12p3Check.h"
 #include "Rule14p4Check.h"
+#include "Rule15p5Check.h"
 #include "Rule15p7Check.h"
 #include "Rule16p7Check.h"
 #include "Rule18p7Check.h"
@@ -50,6 +51,8 @@ public:
         "misrac-rule-12p3");
     CheckFactories.registerCheck<Rule14p4Check>(
         "misrac-rule-14p4");
+    CheckFactories.registerCheck<Rule15p5Check>(
+        "misrac-rule-15p5");
     CheckFactories.registerCheck<Rule15p7Check>(
         "misrac-rule-15p7");
     CheckFactories.registerCheck<Rule16p7Check>(